+++ UI/FormTransferOrder/FormTransferOrder.cs
@@ -25,7 +25,7 @@ namespace WMS.UI.FormTransferOrder
                 MessageBox.Show("请选择一项移库单查看物料条目！", "提示", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 return;
             }
-            var rowData = this.model1.GetRows(new long[] { this.model1.SelectionRange.Row })[0];
+            var rowData = this.model1.GetRows(new int[] { this.model1.SelectionRange.Row })[0];
             new FormTransferOrderItem(rowData).Show();
         }
